working code but not looping
This code works exactly as intended -- the animation plays on both the
transmitting and receiving microbits, and correctly changes between the
two.

However, it doesn't loop. Adding a while loop created animation issues.
I wonder if animations always have to finish on the microbit? How can I
interrupt the animatioon and not have it resume?