animation still short, players to reset
In this version, the animation is still rather short (loops seem to
break the program a lot), and players have to consciously press B to
stop the animation/reset to the default white LED standby mode. This is
possibly a beature, as players consciously move forward together?
Perhaps.

But it's really annoying to not just be able to tell the animation to
loop until it's interrupted without things breaking.